Implement more settings relating to the camera

+++ src/gui/input.cpp
@@ -37,7 +37,7 @@ namespace Toolbox::UI::Input {
 
         void SetMouseState(uint32_t button, bool pressed) { mMouseButtonsDown[button] = pressed; }
 
-        void SetMousePosition(uint32_t x, uint32_t y) { mMousePosition = ImVec2(x, y); }
+        void SetMousePosition(uint32_t x, uint32_t y) { mMousePosition = ImVec2((float)x, (float)y); }
 
         void SetMouseScrollDelta(uint32_t delta) { mMouseScrollDelta = delta; }
     }  // namespace
@@ -75,7 +75,7 @@ void Toolbox::UI::Input::SetMousePosition(ImVec2 pos, bool overwrite_delta) {
         SetMouseWrapped(true);
     }
 #if WIN32
-    SetCursorPos(pos.x, pos.y);
+    SetCursorPos((int)pos.x, (int)pos.y);
 #elif __linux__
     Display *display = XOpenDisplay(0);
     if (display == nullptr)
@@ -127,7 +127,7 @@ void Toolbox::UI::Input::GLFWKeyCallback(GLFWwindow *window, int key, int scanco
 }
 
 void Toolbox::UI::Input::GLFWMousePositionCallback(GLFWwindow *window, double xpos, double ypos) {
-    SetMousePosition(xpos, ypos);
+    SetMousePosition((uint32_t)xpos, (uint32_t)ypos);
 
     ImGui_ImplGlfw_CursorPosCallback(window, xpos, ypos);
 }
@@ -147,7 +147,7 @@ void Toolbox::UI::Input::GLFWMouseButtonCallback(GLFWwindow *window, int button,
 
 void Toolbox::UI::Input::GLFWMouseScrollCallback(GLFWwindow *window, double xoffset,
                                                  double yoffset) {
-    SetMouseScrollDelta(yoffset);
+    SetMouseScrollDelta((uint32_t)yoffset);
 
     ImGui_ImplGlfw_ScrollCallback(window, xoffset, yoffset);
 }
